Project structure and minimal articles.
Articles are going to be big!  So big!  Etc.  But the truth is we don't
need much of that in order to prove a concept.  This commit has a
ridiculously small, memory-persistence-only, article model.  That way we
can see what using that model looks like.

Next up, I think, is standing up an express api, the service, the service
client, and the cli in that order.